Move format_size to pip._internal.utils.ui
Why: This removes the only dependency of utils.ui on utils.misc
